feat: move logging (error)
