+++ cmd/engine/main.go
@@ -13,9 +13,9 @@ import (
 	flag "github.com/spf13/pflag"
 	"github.com/spf13/viper"
 
-	"github.com/promiseofcake/artifactsmmo-engine/cmd/logging"
 	"github.com/promiseofcake/artifactsmmo-engine/internal/actions"
 	"github.com/promiseofcake/artifactsmmo-engine/internal/engine"
+	"github.com/promiseofcake/artifactsmmo-engine/internal/logging"
 )
 
 func init() {
