+++ internal/engine/fight.go
@@ -5,8 +5,8 @@ import (
 	"math"
 	"time"
 
-	"github.com/promiseofcake/artifactsmmo-engine/cmd/logging"
 	"github.com/promiseofcake/artifactsmmo-engine/internal/actions"
+	"github.com/promiseofcake/artifactsmmo-engine/internal/logging"
 	"github.com/promiseofcake/artifactsmmo-engine/internal/models"
 
 	"github.com/promiseofcake/artifactsmmo-go-client/client"
